refactor: Move process handling out of task handler WIP
